+++ src/singleTable/model/from/fromCollection/definitions/index.ts
@@ -1,0 +1,2 @@
+export type { FromCollection, FromCollectionMethods } from './constructor';
+export type { GetCollectionParams, GetCollectionResult } from './get';