+++ src/singleTable/model/from/fromEntity/definitions/list.ts
@@ -1,5 +1,5 @@
-import { AnyObject } from 'types';
-import { BasicRangeKeyConfig, BetweenRangeKeyConfig, QueryParams } from 'provider/utils';
+import type { AnyObject } from 'types';
+import type { BasicRangeKeyConfig, BetweenRangeKeyConfig, QueryParams } from 'provider/utils';
 
 type BasicRangeConfig = {
   value: string;