+++ src/singleTable/model/index.ts
@@ -1,2 +1,10 @@
 export * from './definitions';
 export * from './schema';
+export * from './schemaType';
+
+export {
+  type FromCollection,
+  type FromEntity,
+  type GetCollectionParams,
+  type GetCollectionResult,
+} from './from';